Change the max_delay parameter by an int

+++ python_async_function/2-measure_runtime.py
@@ -14,7 +14,7 @@ def measure_time(n: int, max_delay: int) -> float:
     """
     Params:
         n: int
-        max_delay: float
+        max_delay: int
     Return: a float: total_time / n
     """
